Struct rusoto_autoscaling::AutoScalingGroup

pub struct AutoScalingGroup {
    pub auto_scaling_group_arn: Option<String>,
    pub auto_scaling_group_name: String,
    pub availability_zones: Vec<String>,
    pub created_time: String,
    pub default_cooldown: i64,
    pub desired_capacity: i64,
    pub enabled_metrics: Option<Vec<EnabledMetric>>,
    pub health_check_grace_period: Option<i64>,
    pub health_check_type: String,
    pub instances: Option<Vec<Instance>>,
    pub launch_configuration_name: Option<String>,
    pub load_balancer_names: Option<Vec<String>>,
    pub max_size: i64,
    pub min_size: i64,
    pub new_instances_protected_from_scale_in: Option<bool>,
    pub placement_group: Option<String>,
    pub status: Option<String>,
    pub suspended_processes: Option<Vec<SuspendedProcess>>,
    pub tags: Option<Vec<TagDescription>>,
    pub target_group_ar_ns: Option<Vec<String>>,
    pub termination_policies: Option<Vec<String>>,
    pub vpc_zone_identifier: Option<String>,
}

Describes an Auto Scaling group.

Fields
auto_scaling_group_arn: Option<String>

The Amazon Resource Name (ARN) of the group.

auto_scaling_group_name: String

The name of the group.

availability_zones: Vec<String>

One or more Availability Zones for the group.

created_time: String

The date and time the group was created.

default_cooldown: i64

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start.

desired_capacity: i64

The desired size of the group.

enabled_metrics: Option<Vec<EnabledMetric>>

The metrics enabled for the group.

health_check_grace_period: Option<i64>

The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

health_check_type: String

The service to use for the health checks. The valid values are EC2 and ELB.

instances: Option<Vec<Instance>>

The EC2 instances associated with the group.

launch_configuration_name: Option<String>

The name of the associated launch configuration.

load_balancer_names: Option<Vec<String>>

One or more load balancers associated with the group.

max_size: i64

The maximum size of the group.

min_size: i64

The minimum size of the group.

new_instances_protected_from_scale_in: Option<bool>

Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

placement_group: Option<String>

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide.

status: Option<String>

The current state of the group when DeleteAutoScalingGroup is in progress.

suspended_processes: Option<Vec<SuspendedProcess>>

The suspended processes associated with the group.

tags: Option<Vec<TagDescription>>

The tags for the group.

target_group_ar_ns: Option<Vec<String>>

The Amazon Resource Names (ARN) of the target groups for your load balancer.

termination_policies: Option<Vec<String>>

The termination policies for the group.

vpc_zone_identifier: Option<String>

One or more subnet IDs, if applicable, separated by commas.

If you specify VPCZoneIdentifier and AvailabilityZones, ensure that the Availability Zones of the subnets match the values for AvailabilityZones.