Struct rusoto_autoscaling::LifecycleHook

pub struct LifecycleHook {
    pub auto_scaling_group_name: Option<String>,
    pub default_result: Option<String>,
    pub global_timeout: Option<i64>,
    pub heartbeat_timeout: Option<i64>,
    pub lifecycle_hook_name: Option<String>,
    pub lifecycle_transition: Option<String>,
    pub notification_metadata: Option<String>,
    pub notification_target_arn: Option<String>,
    pub role_arn: Option<String>,
}

Describes a lifecycle hook, which tells Auto Scaling that you want to perform an action when an instance launches or terminates. When you have a lifecycle hook in place, the Auto Scaling group will either:

Pause the instance after it launches, but before it is put into service

Pause the instance as it terminates, but before it is fully terminated

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide.

Fields
auto_scaling_group_name: Option<String>

The name of the Auto Scaling group for the lifecycle hook.

default_result: Option<String>

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON. The default value is CONTINUE.

global_timeout: Option<i64>

The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout, whichever is smaller.

heartbeat_timeout: Option<i64>

The maximum time, in seconds, that can elapse before the lifecycle hook times out. The default is 3600 seconds (1 hour). When the lifecycle hook times out, Auto Scaling performs the default action. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

lifecycle_hook_name: Option<String>

The name of the lifecycle hook.

lifecycle_transition: Option<String>

The state of the EC2 instance to which you want to attach the lifecycle hook. For a list of lifecycle hook types, see DescribeLifecycleHookTypes.

notification_metadata: Option<String>

Additional information that you want to include any time Auto Scaling sends a message to the notification target.

notification_target_arn: Option<String>

The ARN of the notification target that Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target includes the following:

Lifecycle action token

User account ID

Name of the Auto Scaling group

Lifecycle hook name

EC2 instance ID

Lifecycle transition

Notification metadata

role_arn: Option<String>

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.