Struct rusoto_autoscaling::PutScheduledUpdateGroupActionType

pub struct PutScheduledUpdateGroupActionType {
    pub auto_scaling_group_name: String,
    pub desired_capacity: Option<i64>,
    pub end_time: Option<String>,
    pub max_size: Option<i64>,
    pub min_size: Option<i64>,
    pub recurrence: Option<String>,
    pub scheduled_action_name: String,
    pub start_time: Option<String>,
    pub time: Option<String>,
}

Fields
auto_scaling_group_name: String

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

desired_capacity: Option<i64>

The number of EC2 instances that should be running in the group.

end_time: Option<String>

The time for the recurring schedule to end. Auto Scaling does not perform the action after this time.

max_size: Option<i64>

The maximum size for the Auto Scaling group.

min_size: Option<i64>

The minimum size for the Auto Scaling group.

recurrence: Option<String>

The recurring schedule for this action, in Unix cron syntax format. For more information, see Cron in Wikipedia.

scheduled_action_name: String

The name of this scaling action.

start_time: Option<String>

The time for this action to start, in "YYYY-MM-DDThh:mm:ssZ" format in UTC/GMT only (for example, 2014-06-01T00:00:00Z).

If you specify Recurrence and StartTime, Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule your action in the past, Auto Scaling returns an error message.

time: Option<String>

This parameter is deprecated.