Struct rusoto_autoscaling::ScalingPolicy

pub struct ScalingPolicy {
    pub adjustment_type: Option<String>,
    pub alarms: Option<Vec<Alarm>>,
    pub auto_scaling_group_name: Option<String>,
    pub cooldown: Option<i64>,
    pub estimated_instance_warmup: Option<i64>,
    pub metric_aggregation_type: Option<String>,
    pub min_adjustment_magnitude: Option<i64>,
    pub min_adjustment_step: Option<i64>,
    pub policy_arn: Option<String>,
    pub policy_name: Option<String>,
    pub policy_type: Option<String>,
    pub scaling_adjustment: Option<i64>,
    pub step_adjustments: Option<Vec<StepAdjustment>>,
}

Describes a scaling policy.

Fields
adjustment_type: Option<String>

The adjustment type, which specifies how ScalingAdjustment is interpreted. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

alarms: Option<Vec<Alarm>>

The CloudWatch alarms related to the policy.

auto_scaling_group_name: Option<String>

The name of the Auto Scaling group associated with this scaling policy.

cooldown: Option<i64>

The amount of time, in seconds, after a scaling activity completes before any further dynamic scaling activities can start.

estimated_instance_warmup: Option<i64>

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

metric_aggregation_type: Option<String>

The aggregation type for the CloudWatch metrics. Valid values are Minimum, Maximum, and Average.

min_adjustment_magnitude: Option<i64>

The minimum number of instances to scale. If the value of AdjustmentType is PercentChangeInCapacity, the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least this many instances. Otherwise, the error is ValidationError.

min_adjustment_step: Option<i64>

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

policy_arn: Option<String>

The Amazon Resource Name (ARN) of the policy.

policy_name: Option<String>

The name of the scaling policy.

policy_type: Option<String>

The policy type. Valid values are SimpleScaling and StepScaling.

scaling_adjustment: Option<i64>

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

step_adjustments: Option<Vec<StepAdjustment>>

A set of adjustments that enable you to scale based on the size of the alarm breach.