Crate rusoto_autoscaling

Amazon EC2 Auto Scaling
Amazon EC2 Auto Scaling is designed to automatically launch or terminate EC2 instances based on user-defined policies, schedules, and health checks. Use this service in conjunction with the AWS Auto Scaling, Amazon CloudWatch, and Elastic Load Balancing services.

If you're using the service, you're probably looking for AutoscalingClient and Autoscaling.

Structs

ActivitiesType

Activity
Describes scaling activity, which is a long-running process that represents a change to your Auto Scaling group, such as changing its size or replacing an instance.

ActivityType

AdjustmentType
Describes a policy adjustment type.

For more information, see Dynamic Scaling in the Auto Scaling User Guide.

Alarm
Describes an alarm.

AttachInstancesQuery

AttachLoadBalancerTargetGroupsResultType

AttachLoadBalancerTargetGroupsType

AttachLoadBalancersResultType

AttachLoadBalancersType

AutoScalingGroup
Describes an Auto Scaling group.

AutoScalingGroupNamesType

AutoScalingGroupsType

AutoScalingInstanceDetails
Describes an EC2 instance associated with an Auto Scaling group.

AutoScalingInstancesType

AutoscalingClient
A client for the Auto Scaling API.

BlockDeviceMapping
Describes a block device mapping.

CompleteLifecycleActionAnswer

CompleteLifecycleActionType

CreateAutoScalingGroupType

CreateLaunchConfigurationType

CreateOrUpdateTagsType

CustomizedMetricSpecification
Configures a customized metric for a target tracking policy.

DeleteAutoScalingGroupType

DeleteLifecycleHookAnswer

DeleteLifecycleHookType

DeleteNotificationConfigurationType

DeletePolicyType

DeleteScheduledActionType

DeleteTagsType

DescribeAccountLimitsAnswer

DescribeAdjustmentTypesAnswer

DescribeAutoScalingInstancesType

DescribeAutoScalingNotificationTypesAnswer

DescribeLifecycleHookTypesAnswer

DescribeLifecycleHooksAnswer

DescribeLifecycleHooksType

DescribeLoadBalancerTargetGroupsRequest

DescribeLoadBalancerTargetGroupsResponse

DescribeLoadBalancersRequest

DescribeLoadBalancersResponse

DescribeMetricCollectionTypesAnswer

DescribeNotificationConfigurationsAnswer

DescribeNotificationConfigurationsType

DescribePoliciesType

DescribeScalingActivitiesType

DescribeScheduledActionsType

DescribeTagsType

DescribeTerminationPolicyTypesAnswer

DetachInstancesAnswer

DetachInstancesQuery

DetachLoadBalancerTargetGroupsResultType

DetachLoadBalancerTargetGroupsType

DetachLoadBalancersResultType

DetachLoadBalancersType

DisableMetricsCollectionQuery

Ebs
Describes an Amazon EBS volume.

EnableMetricsCollectionQuery

EnabledMetric
Describes an enabled metric.

EnterStandbyAnswer

EnterStandbyQuery

ExecutePolicyType

ExitStandbyAnswer

ExitStandbyQuery

Filter
Describes a filter.

Instance
Describes an EC2 instance.

InstanceMonitoring
Describes whether detailed monitoring is enabled for the Auto Scaling instances.

LaunchConfiguration
Describes a launch configuration.

LaunchConfigurationNameType

LaunchConfigurationNamesType

LaunchConfigurationsType

LaunchTemplateSpecification
Describes a launch template.

LifecycleHook
Describes a lifecycle hook, which tells Auto Scaling that you want to perform an action whenever it launches instances or whenever it terminates instances.

For more information, see Auto Scaling Lifecycle Hooks in the Auto Scaling User Guide.

LifecycleHookSpecification
Describes a lifecycle hook, which tells Auto Scaling that you want to perform an action whenever it launches instances or whenever it terminates instances.

For more information, see Auto Scaling Lifecycle Hooks in the Auto Scaling User Guide.

LoadBalancerState
Describes the state of a Classic Load Balancer.

If you specify a load balancer when creating the Auto Scaling group, the state of the load balancer is InService.

If you attach a load balancer to an existing Auto Scaling group, the initial state is Adding. The state transitions to Added after all instances in the group are registered with the load balancer. If ELB health checks are enabled for the load balancer, the state transitions to InService after at least one instance in the group passes the health check. If EC2 health checks are enabled instead, the load balancer remains in the Added state.

LoadBalancerTargetGroupState
Describes the state of a target group.

If you attach a target group to an existing Auto Scaling group, the initial state is Adding. The state transitions to Added after all Auto Scaling instances are registered with the target group. If ELB health checks are enabled, the state transitions to InService after at least one Auto Scaling instance passes the health check. If EC2 health checks are enabled instead, the target group remains in the Added state.

MetricCollectionType
Describes a metric.

MetricDimension
Describes the dimension of a metric.

MetricGranularityType
Describes a granularity of a metric.

NotificationConfiguration
Describes a notification.

PoliciesType

PolicyARNType
Contains the output of PutScalingPolicy.

PredefinedMetricSpecification
Configures a predefined metric for a target tracking policy.

ProcessType
Describes a process type.

For more information, see Auto Scaling Processes in the Auto Scaling User Guide.

ProcessesType

PutLifecycleHookAnswer

PutLifecycleHookType

PutNotificationConfigurationType

PutScalingPolicyType

PutScheduledUpdateGroupActionType

RecordLifecycleActionHeartbeatAnswer

RecordLifecycleActionHeartbeatType

ScalingPolicy
Describes a scaling policy.

ScalingProcessQuery

ScheduledActionsType

ScheduledUpdateGroupAction
Describes a scheduled update to an Auto Scaling group.

SetDesiredCapacityType

SetInstanceHealthQuery

SetInstanceProtectionAnswer

SetInstanceProtectionQuery

StepAdjustment

Describes an adjustment based on the difference between the value of the aggregated CloudWatch metric and the breach threshold that you've defined for the alarm.

For the following examples, suppose that you have an alarm with a breach threshold of 50:

If you want the adjustment to be triggered when the metric is greater than or equal to 50 and less than 60, specify a lower bound of 0 and an upper bound of 10.

If you want the adjustment to be triggered when the metric is greater than 40 and less than or equal to 50, specify a lower bound of -10 and an upper bound of 0.

There are a few rules for the step adjustments for your step policy:

The ranges of your step adjustments can't overlap or have a gap.

At most one step adjustment can have a null lower bound. If one step adjustment has a negative lower bound, then there must be a step adjustment with a null lower bound.

At most one step adjustment can have a null upper bound. If one step adjustment has a positive upper bound, then there must be a step adjustment with a null upper bound.

The upper and lower bound can't be null in the same step adjustment.

SuspendedProcess
Describes an Auto Scaling process that has been suspended. For more information, see ProcessType.

Tag
Describes a tag for an Auto Scaling group.

TagDescription
Describes a tag for an Auto Scaling group.

TagsType

TargetTrackingConfiguration
Represents a target tracking policy configuration.

TerminateInstanceInAutoScalingGroupType

UpdateAutoScalingGroupType

Enums

AttachInstancesError
Errors returned by AttachInstances

AttachLoadBalancerTargetGroupsError
Errors returned by AttachLoadBalancerTargetGroups

AttachLoadBalancersError
Errors returned by AttachLoadBalancers

CompleteLifecycleActionError
Errors returned by CompleteLifecycleAction

CreateAutoScalingGroupError
Errors returned by CreateAutoScalingGroup

CreateLaunchConfigurationError
Errors returned by CreateLaunchConfiguration

CreateOrUpdateTagsError
Errors returned by CreateOrUpdateTags

DeleteAutoScalingGroupError
Errors returned by DeleteAutoScalingGroup

DeleteLaunchConfigurationError
Errors returned by DeleteLaunchConfiguration

DeleteLifecycleHookError
Errors returned by DeleteLifecycleHook

DeleteNotificationConfigurationError
Errors returned by DeleteNotificationConfiguration

DeletePolicyError
Errors returned by DeletePolicy

DeleteScheduledActionError
Errors returned by DeleteScheduledAction

DeleteTagsError
Errors returned by DeleteTags

DescribeAccountLimitsError
Errors returned by DescribeAccountLimits

DescribeAdjustmentTypesError
Errors returned by DescribeAdjustmentTypes

DescribeAutoScalingGroupsError
Errors returned by DescribeAutoScalingGroups

DescribeAutoScalingInstancesError
Errors returned by DescribeAutoScalingInstances

DescribeAutoScalingNotificationTypesError
Errors returned by DescribeAutoScalingNotificationTypes

DescribeLaunchConfigurationsError
Errors returned by DescribeLaunchConfigurations

DescribeLifecycleHookTypesError
Errors returned by DescribeLifecycleHookTypes

DescribeLifecycleHooksError
Errors returned by DescribeLifecycleHooks

DescribeLoadBalancerTargetGroupsError
Errors returned by DescribeLoadBalancerTargetGroups

DescribeLoadBalancersError
Errors returned by DescribeLoadBalancers

DescribeMetricCollectionTypesError
Errors returned by DescribeMetricCollectionTypes

DescribeNotificationConfigurationsError
Errors returned by DescribeNotificationConfigurations

DescribePoliciesError
Errors returned by DescribePolicies

DescribeScalingActivitiesError
Errors returned by DescribeScalingActivities

DescribeScalingProcessTypesError
Errors returned by DescribeScalingProcessTypes

DescribeScheduledActionsError
Errors returned by DescribeScheduledActions

DescribeTagsError
Errors returned by DescribeTags

DescribeTerminationPolicyTypesError
Errors returned by DescribeTerminationPolicyTypes

DetachInstancesError
Errors returned by DetachInstances

DetachLoadBalancerTargetGroupsError
Errors returned by DetachLoadBalancerTargetGroups

DetachLoadBalancersError
Errors returned by DetachLoadBalancers

DisableMetricsCollectionError
Errors returned by DisableMetricsCollection

EnableMetricsCollectionError
Errors returned by EnableMetricsCollection

EnterStandbyError
Errors returned by EnterStandby

ExecutePolicyError
Errors returned by ExecutePolicy

ExitStandbyError
Errors returned by ExitStandby

PutLifecycleHookError
Errors returned by PutLifecycleHook

PutNotificationConfigurationError
Errors returned by PutNotificationConfiguration

PutScalingPolicyError
Errors returned by PutScalingPolicy

PutScheduledUpdateGroupActionError
Errors returned by PutScheduledUpdateGroupAction

RecordLifecycleActionHeartbeatError
Errors returned by RecordLifecycleActionHeartbeat

ResumeProcessesError
Errors returned by ResumeProcesses

SetDesiredCapacityError
Errors returned by SetDesiredCapacity

SetInstanceHealthError
Errors returned by SetInstanceHealth

SetInstanceProtectionError
Errors returned by SetInstanceProtection

SuspendProcessesError
Errors returned by SuspendProcesses

TerminateInstanceInAutoScalingGroupError
Errors returned by TerminateInstanceInAutoScalingGroup

UpdateAutoScalingGroupError
Errors returned by UpdateAutoScalingGroup

Traits

Autoscaling
Trait representing the capabilities of the Auto Scaling API. Auto Scaling clients implement this trait.