Struct rusoto_autoscaling::AutoscalingClient

pub struct AutoscalingClient { /* fields omitted */ }

A client for the Auto Scaling API.

Methods

impl AutoscalingClient

pub fn new(region: Region) -> AutoscalingClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AutoscalingClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,