Struct rusoto_autoscaling::CreateAutoScalingGroupType

pub struct CreateAutoScalingGroupType {
    pub auto_scaling_group_name: String,
    pub availability_zones: Option<Vec<String>>,
    pub default_cooldown: Option<i64>,
    pub desired_capacity: Option<i64>,
    pub health_check_grace_period: Option<i64>,
    pub health_check_type: Option<String>,
    pub instance_id: Option<String>,
    pub launch_configuration_name: Option<String>,
    pub launch_template: Option<LaunchTemplateSpecification>,
    pub lifecycle_hook_specification_list: Option<Vec<LifecycleHookSpecification>>,
    pub load_balancer_names: Option<Vec<String>>,
    pub max_size: i64,
    pub min_size: i64,
    pub mixed_instances_policy: Option<MixedInstancesPolicy>,
    pub new_instances_protected_from_scale_in: Option<bool>,
    pub placement_group: Option<String>,
    pub service_linked_role_arn: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub target_group_ar_ns: Option<Vec<String>>,
    pub termination_policies: Option<Vec<String>>,
    pub vpc_zone_identifier: Option<String>,
}

Fields
auto_scaling_group_name: String

The name of the Auto Scaling group. This name must be unique within the scope of your AWS account.

availability_zones: Option<Vec<String>>

One or more Availability Zones for the group. This parameter is optional if you specify one or more subnets for VPCZoneIdentifier.

Conditional: If your account supports EC2-Classic and VPC, this parameter is required to launch instances into EC2-Classic.

default_cooldown: Option<i64>

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. The default value is 300.

For more information, see Scaling Cooldowns in the Amazon EC2 Auto Scaling User Guide.

desired_capacity: Option<i64>

The number of EC2 instances that should be running in the group. This number must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group. If you do not specify a desired capacity, the default is the minimum size of the group.

health_check_grace_period: Option<i64>

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service. During this time, any health check failures for the instance are ignored. The default value is 0.

For more information, see Health Checks for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

Conditional: This parameter is required if you are adding an ELB health check.

health_check_type: Option<String>

The service to use for the health checks. The valid values are EC2 and ELB. The default value is EC2. If you configure an Auto Scaling group to use ELB health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks.

For more information, see Health Checks for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

instance_id: Option<String>

The ID of the instance used to create a launch configuration for the group. This parameter, a launch configuration, a launch template, or a mixed instances policy must be specified.

When you specify an ID of an instance, Amazon EC2 Auto Scaling creates a new launch configuration and associates it with the group. This launch configuration derives its attributes from the specified instance, except for the block device mapping.

For more information, see Create an Auto Scaling Group Using an EC2 Instance in the Amazon EC2 Auto Scaling User Guide.

launch_configuration_name: Option<String>

The name of the launch configuration. This parameter, a launch template, a mixed instances policy, or an EC2 instance must be specified.

For more information, see Creating an Auto Scaling Group Using a Launch Configuration in the Amazon EC2 Auto Scaling User Guide.

launch_template: Option<LaunchTemplateSpecification>

The launch template to use to launch instances. This parameter, a launch configuration, a mixed instances policy, or an EC2 instance must be specified.

For more information, see Creating an Auto Scaling Group Using a Launch Template in the Amazon EC2 Auto Scaling User Guide.

lifecycle_hook_specification_list: Option<Vec<LifecycleHookSpecification>>

One or more lifecycle hooks.

load_balancer_names: Option<Vec<String>>

One or more Classic Load Balancers. To specify an Application Load Balancer or a Network Load Balancer, use TargetGroupARNs instead.

For more information, see Using a Load Balancer With an Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

max_size: i64

The maximum size of the group.

min_size: i64

The minimum size of the group.

mixed_instances_policy: Option<MixedInstancesPolicy>

The mixed instances policy to use to launch instances. This parameter, a launch template, a launch configuration, or an EC2 instance must be specified.

For more information, see Auto Scaling Groups with Multiple Instance Types and Purchase Options in the Amazon EC2 Auto Scaling User Guide.

new_instances_protected_from_scale_in: Option<bool>

Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in.

For more information about preventing instances from terminating on scale in, see Instance Protection in the Amazon EC2 Auto Scaling User Guide.

placement_group: Option<String>

The name of the placement group into which to launch your instances, if any. A placement group is a logical grouping of instances within a single Availability Zone. You cannot specify multiple Availability Zones and a placement group. For more information, see Placement Groups in the Amazon EC2 User Guide for Linux Instances.

service_linked_role_arn: Option<String>

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other AWS services on your behalf. By default, Amazon EC2 Auto Scaling uses a service-linked role named AWSServiceRoleForAutoScaling, which it creates if it does not exist. For more information, see Service-Linked Roles in the Amazon EC2 Auto Scaling User Guide.

tags: Option<Vec<Tag>>

One or more tags.

For more information, see Tagging Auto Scaling Groups and Instances in the Amazon EC2 Auto Scaling User Guide.

target_group_ar_ns: Option<Vec<String>>

The Amazon Resource Names (ARN) of the target groups.

termination_policies: Option<Vec<String>>

One or more termination policies used to select the instance to terminate. These policies are executed in the order that they are listed.

For more information, see Controlling Which Instances Auto Scaling Terminates During Scale In in the Amazon EC2 Auto Scaling User Guide.

vpc_zone_identifier: Option<String>

A comma-separated list of subnet IDs for your virtual private cloud (VPC).

If you specify VPCZoneIdentifier with AvailabilityZones, the subnets that you specify for this parameter must reside in those Availability Zones.

Conditional: If your account supports EC2-Classic and VPC, this parameter is required to launch instances into a VPC.