Struct rusoto_autoscaling::CreateAutoScalingGroupType

pub struct CreateAutoScalingGroupType {
    pub auto_scaling_group_name: String,
    pub availability_zones: Option<Vec<String>>,
    pub default_cooldown: Option<i64>,
    pub desired_capacity: Option<i64>,
    pub health_check_grace_period: Option<i64>,
    pub health_check_type: Option<String>,
    pub instance_id: Option<String>,
    pub launch_configuration_name: Option<String>,
    pub launch_template: Option<LaunchTemplateSpecification>,
    pub lifecycle_hook_specification_list: Option<Vec<LifecycleHookSpecification>>,
    pub load_balancer_names: Option<Vec<String>>,
    pub max_size: i64,
    pub min_size: i64,
    pub new_instances_protected_from_scale_in: Option<bool>,
    pub placement_group: Option<String>,
    pub service_linked_role_arn: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub target_group_ar_ns: Option<Vec<String>>,
    pub termination_policies: Option<Vec<String>>,
    pub vpc_zone_identifier: Option<String>,
}

Fields
auto_scaling_group_name: String

The name of the Auto Scaling group. This name must be unique within the scope of your AWS account.

availability_zones: Option<Vec<String>>

One or more Availability Zones for the group. This parameter is optional if you specify one or more subnets.

default_cooldown: Option<i64>

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. The default is 300.

For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide.

desired_capacity: Option<i64>

The number of EC2 instances that should be running in the group. This number must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group. If you do not specify a desired capacity, the default is the minimum size of the group.

health_check_grace_period: Option<i64>

The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service. During this time, any health check failures for the instance are ignored. The default is 0.

This parameter is required if you are adding an ELB health check.

For more information, see Health Checks in the Auto Scaling User Guide.

health_check_type: Option<String>

The service to use for the health checks. The valid values are EC2 and ELB.

By default, health checks use Amazon EC2 instance status checks to determine the health of an instance. For more information, see Health Checks in the Auto Scaling User Guide.

instance_id: Option<String>

The ID of the instance used to create a launch configuration for the group. You must specify one of the following: an EC2 instance, a launch configuration, or a launch template.

When you specify an ID of an instance, Auto Scaling creates a new launch configuration and associates it with the group. This launch configuration derives its attributes from the specified instance, with the exception of the block device mapping.

For more information, see Create an Auto Scaling Group Using an EC2 Instance in the Auto Scaling User Guide.

launch_configuration_name: Option<String>

The name of the launch configuration. You must specify one of the following: a launch configuration, a launch template, or an EC2 instance.

launch_template: Option<LaunchTemplateSpecification>

The launch template to use to launch instances. You must specify one of the following: a launch template, a launch configuration, or an EC2 instance.

lifecycle_hook_specification_list: Option<Vec<LifecycleHookSpecification>>

One or more lifecycle hooks.

load_balancer_names: Option<Vec<String>>

One or more Classic Load Balancers. To specify an Application Load Balancer, use TargetGroupARNs instead.

For more information, see Using a Load Balancer With an Auto Scaling Group in the Auto Scaling User Guide.

max_size: i64

The maximum size of the group.

min_size: i64

The minimum size of the group.

new_instances_protected_from_scale_in: Option<bool>

Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

placement_group: Option<String>

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide.

service_linked_role_arn: Option<String>

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other AWS services on your behalf. By default, Auto Scaling uses a service-linked role named AWSServiceRoleForAutoScaling, which it creates if it does not exist.

tags: Option<Vec<Tag>>

One or more tags.

For more information, see Tagging Auto Scaling Groups and Instances in the Auto Scaling User Guide.

target_group_ar_ns: Option<Vec<String>>

The Amazon Resource Names (ARN) of the target groups.

termination_policies: Option<Vec<String>>

One or more termination policies used to select the instance to terminate. These policies are executed in the order that they are listed.

For more information, see Controlling Which Instances Auto Scaling Terminates During Scale In in the Auto Scaling User Guide.

vpc_zone_identifier: Option<String>

A comma-separated list of subnet identifiers for your virtual private cloud (VPC).

If you specify subnets and Availability Zones with this call, ensure that the subnets' Availability Zones match the Availability Zones specified.

For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide.