Trait rusoto_autoscaling::Autoscaling

pub trait Autoscaling {
    fn attach_instances(
        &self, 
        input: AttachInstancesQuery
    ) -> RusotoFuture<(), AttachInstancesError>;

    fn attach_load_balancer_target_groups(
        &self, 
        input: AttachLoadBalancerTargetGroupsType
    ) -> RusotoFuture<AttachLoadBalancerTargetGroupsResultType, AttachLoadBalancerTargetGroupsError>;

    fn attach_load_balancers(
        &self, 
        input: AttachLoadBalancersType
    ) -> RusotoFuture<AttachLoadBalancersResultType, AttachLoadBalancersError>;

    fn batch_delete_scheduled_action(
        &self, 
        input: BatchDeleteScheduledActionType
    ) -> RusotoFuture<BatchDeleteScheduledActionAnswer, BatchDeleteScheduledActionError>;

    fn batch_put_scheduled_update_group_action(
        &self, 
        input: BatchPutScheduledUpdateGroupActionType
    ) -> RusotoFuture<BatchPutScheduledUpdateGroupActionAnswer, BatchPutScheduledUpdateGroupActionError>;

    fn complete_lifecycle_action(
        &self, 
        input: CompleteLifecycleActionType
    ) -> RusotoFuture<CompleteLifecycleActionAnswer, CompleteLifecycleActionError>;

    fn create_auto_scaling_group(
        &self, 
        input: CreateAutoScalingGroupType
    ) -> RusotoFuture<(), CreateAutoScalingGroupError>;

    fn create_launch_configuration(
        &self, 
        input: CreateLaunchConfigurationType
    ) -> RusotoFuture<(), CreateLaunchConfigurationError>;

    fn create_or_update_tags(
        &self, 
        input: CreateOrUpdateTagsType
    ) -> RusotoFuture<(), CreateOrUpdateTagsError>;

    fn delete_auto_scaling_group(
        &self, 
        input: DeleteAutoScalingGroupType
    ) -> RusotoFuture<(), DeleteAutoScalingGroupError>;

    fn delete_launch_configuration(
        &self, 
        input: LaunchConfigurationNameType
    ) -> RusotoFuture<(), DeleteLaunchConfigurationError>;

    fn delete_lifecycle_hook(
        &self, 
        input: DeleteLifecycleHookType
    ) -> RusotoFuture<DeleteLifecycleHookAnswer, DeleteLifecycleHookError>;

    fn delete_notification_configuration(
        &self, 
        input: DeleteNotificationConfigurationType
    ) -> RusotoFuture<(), DeleteNotificationConfigurationError>;

    fn delete_policy(
        &self, 
        input: DeletePolicyType
    ) -> RusotoFuture<(), DeletePolicyError>;

    fn delete_scheduled_action(
        &self, 
        input: DeleteScheduledActionType
    ) -> RusotoFuture<(), DeleteScheduledActionError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsType
    ) -> RusotoFuture<(), DeleteTagsError>;

    fn describe_account_limits(
        &self
    ) -> RusotoFuture<DescribeAccountLimitsAnswer, DescribeAccountLimitsError>;

    fn describe_adjustment_types(
        &self
    ) -> RusotoFuture<DescribeAdjustmentTypesAnswer, DescribeAdjustmentTypesError>;

    fn describe_auto_scaling_groups(
        &self, 
        input: AutoScalingGroupNamesType
    ) -> RusotoFuture<AutoScalingGroupsType, DescribeAutoScalingGroupsError>;

    fn describe_auto_scaling_instances(
        &self, 
        input: DescribeAutoScalingInstancesType
    ) -> RusotoFuture<AutoScalingInstancesType, DescribeAutoScalingInstancesError>;

    fn describe_auto_scaling_notification_types(
        &self
    ) -> RusotoFuture<DescribeAutoScalingNotificationTypesAnswer, DescribeAutoScalingNotificationTypesError>;

    fn describe_launch_configurations(
        &self, 
        input: LaunchConfigurationNamesType
    ) -> RusotoFuture<LaunchConfigurationsType, DescribeLaunchConfigurationsError>;

    fn describe_lifecycle_hook_types(
        &self
    ) -> RusotoFuture<DescribeLifecycleHookTypesAnswer, DescribeLifecycleHookTypesError>;

    fn describe_lifecycle_hooks(
        &self, 
        input: DescribeLifecycleHooksType
    ) -> RusotoFuture<DescribeLifecycleHooksAnswer, DescribeLifecycleHooksError>;

    fn describe_load_balancer_target_groups(
        &self, 
        input: DescribeLoadBalancerTargetGroupsRequest
    ) -> RusotoFuture<DescribeLoadBalancerTargetGroupsResponse, DescribeLoadBalancerTargetGroupsError>;

    fn describe_load_balancers(
        &self, 
        input: DescribeLoadBalancersRequest
    ) -> RusotoFuture<DescribeLoadBalancersResponse, DescribeLoadBalancersError>;

    fn describe_metric_collection_types(
        &self
    ) -> RusotoFuture<DescribeMetricCollectionTypesAnswer, DescribeMetricCollectionTypesError>;

    fn describe_notification_configurations(
        &self, 
        input: DescribeNotificationConfigurationsType
    ) -> RusotoFuture<DescribeNotificationConfigurationsAnswer, DescribeNotificationConfigurationsError>;

    fn describe_policies(
        &self, 
        input: DescribePoliciesType
    ) -> RusotoFuture<PoliciesType, DescribePoliciesError>;

    fn describe_scaling_activities(
        &self, 
        input: DescribeScalingActivitiesType
    ) -> RusotoFuture<ActivitiesType, DescribeScalingActivitiesError>;

    fn describe_scaling_process_types(
        &self
    ) -> RusotoFuture<ProcessesType, DescribeScalingProcessTypesError>;

    fn describe_scheduled_actions(
        &self, 
        input: DescribeScheduledActionsType
    ) -> RusotoFuture<ScheduledActionsType, DescribeScheduledActionsError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsType
    ) -> RusotoFuture<TagsType, DescribeTagsError>;

    fn describe_termination_policy_types(
        &self
    ) -> RusotoFuture<DescribeTerminationPolicyTypesAnswer, DescribeTerminationPolicyTypesError>;

    fn detach_instances(
        &self, 
        input: DetachInstancesQuery
    ) -> RusotoFuture<DetachInstancesAnswer, DetachInstancesError>;

    fn detach_load_balancer_target_groups(
        &self, 
        input: DetachLoadBalancerTargetGroupsType
    ) -> RusotoFuture<DetachLoadBalancerTargetGroupsResultType, DetachLoadBalancerTargetGroupsError>;

    fn detach_load_balancers(
        &self, 
        input: DetachLoadBalancersType
    ) -> RusotoFuture<DetachLoadBalancersResultType, DetachLoadBalancersError>;

    fn disable_metrics_collection(
        &self, 
        input: DisableMetricsCollectionQuery
    ) -> RusotoFuture<(), DisableMetricsCollectionError>;

    fn enable_metrics_collection(
        &self, 
        input: EnableMetricsCollectionQuery
    ) -> RusotoFuture<(), EnableMetricsCollectionError>;

    fn enter_standby(
        &self, 
        input: EnterStandbyQuery
    ) -> RusotoFuture<EnterStandbyAnswer, EnterStandbyError>;

    fn execute_policy(
        &self, 
        input: ExecutePolicyType
    ) -> RusotoFuture<(), ExecutePolicyError>;

    fn exit_standby(
        &self, 
        input: ExitStandbyQuery
    ) -> RusotoFuture<ExitStandbyAnswer, ExitStandbyError>;

    fn put_lifecycle_hook(
        &self, 
        input: PutLifecycleHookType
    ) -> RusotoFuture<PutLifecycleHookAnswer, PutLifecycleHookError>;

    fn put_notification_configuration(
        &self, 
        input: PutNotificationConfigurationType
    ) -> RusotoFuture<(), PutNotificationConfigurationError>;

    fn put_scaling_policy(
        &self, 
        input: PutScalingPolicyType
    ) -> RusotoFuture<PolicyARNType, PutScalingPolicyError>;

    fn put_scheduled_update_group_action(
        &self, 
        input: PutScheduledUpdateGroupActionType
    ) -> RusotoFuture<(), PutScheduledUpdateGroupActionError>;

    fn record_lifecycle_action_heartbeat(
        &self, 
        input: RecordLifecycleActionHeartbeatType
    ) -> RusotoFuture<RecordLifecycleActionHeartbeatAnswer, RecordLifecycleActionHeartbeatError>;

    fn resume_processes(
        &self, 
        input: ScalingProcessQuery
    ) -> RusotoFuture<(), ResumeProcessesError>;

    fn set_desired_capacity(
        &self, 
        input: SetDesiredCapacityType
    ) -> RusotoFuture<(), SetDesiredCapacityError>;

    fn set_instance_health(
        &self, 
        input: SetInstanceHealthQuery
    ) -> RusotoFuture<(), SetInstanceHealthError>;

    fn set_instance_protection(
        &self, 
        input: SetInstanceProtectionQuery
    ) -> RusotoFuture<SetInstanceProtectionAnswer, SetInstanceProtectionError>;

    fn suspend_processes(
        &self, 
        input: ScalingProcessQuery
    ) -> RusotoFuture<(), SuspendProcessesError>;

    fn terminate_instance_in_auto_scaling_group(
        &self, 
        input: TerminateInstanceInAutoScalingGroupType
    ) -> RusotoFuture<ActivityType, TerminateInstanceInAutoScalingGroupError>;

    fn update_auto_scaling_group(
        &self, 
        input: UpdateAutoScalingGroupType
    ) -> RusotoFuture<(), UpdateAutoScalingGroupError>;
}

Trait representing the capabilities of the Auto Scaling API. Auto Scaling clients implement this trait.

Required methods

fn attach_instances(
&self,
input: AttachInstancesQuery
) -> RusotoFuture<(), AttachInstancesError>

Attaches one or more EC2 instances to the specified Auto Scaling group.

When you attach instances, Amazon EC2 Auto Scaling increases the desired capacity of the group by the number of instances being attached. If the number of instances being attached plus the desired capacity of the group exceeds the maximum size of the group, the operation fails.

If there is a Classic Load Balancer attached to your Auto Scaling group, the instances are also registered with the load balancer. If there are target groups attached to your Auto Scaling group, the instances are also registered with the target groups.

For more information, see Attach EC2 Instances to Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn attach_load_balancer_target_groups(
&self,
input: AttachLoadBalancerTargetGroupsType
) -> RusotoFuture<AttachLoadBalancerTargetGroupsResultType, AttachLoadBalancerTargetGroupsError>

Attaches one or more target groups to the specified Auto Scaling group.

To describe the target groups for an Auto Scaling group, use DescribeLoadBalancerTargetGroups. To detach the target group from the Auto Scaling group, use DetachLoadBalancerTargetGroups.

With Application Load Balancers and Network Load Balancers, instances are registered as targets with a target group. With Classic Load Balancers, instances are registered with the load balancer. For more information, see Attaching a Load Balancer to Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn attach_load_balancers(
&self,
input: AttachLoadBalancersType
) -> RusotoFuture<AttachLoadBalancersResultType, AttachLoadBalancersError>

Attaches one or more Classic Load Balancers to the specified Auto Scaling group.

To attach an Application Load Balancer or a Network Load Balancer instead, see AttachLoadBalancerTargetGroups.

To describe the load balancers for an Auto Scaling group, use DescribeLoadBalancers. To detach the load balancer from the Auto Scaling group, use DetachLoadBalancers.

For more information, see Attaching a Load Balancer to Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn batch_delete_scheduled_action(
&self,
input: BatchDeleteScheduledActionType
) -> RusotoFuture<BatchDeleteScheduledActionAnswer, BatchDeleteScheduledActionError>

Deletes one or more scheduled actions for the specified Auto Scaling group.

fn batch_put_scheduled_update_group_action(
&self,
input: BatchPutScheduledUpdateGroupActionType
) -> RusotoFuture<BatchPutScheduledUpdateGroupActionAnswer, BatchPutScheduledUpdateGroupActionError>

Creates or updates one or more scheduled scaling actions for an Auto Scaling group. If you leave a parameter unspecified when updating a scheduled scaling action, the corresponding value remains unchanged.

fn complete_lifecycle_action(
&self,
input: CompleteLifecycleActionType
) -> RusotoFuture<CompleteLifecycleActionAnswer, CompleteLifecycleActionError>

Completes the lifecycle action for the specified token or instance with the specified result.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Amazon EC2 Auto Scaling Lifecycle Hooks in the Amazon EC2 Auto Scaling User Guide.

fn create_auto_scaling_group(
&self,
input: CreateAutoScalingGroupType
) -> RusotoFuture<(), CreateAutoScalingGroupError>

Creates an Auto Scaling group with the specified name and attributes.

If you exceed your maximum limit of Auto Scaling groups, the call fails. For information about viewing this limit, see DescribeAccountLimits. For information about updating this limit, see Amazon EC2 Auto Scaling Limits in the Amazon EC2 Auto Scaling User Guide.

fn create_launch_configuration(
&self,
input: CreateLaunchConfigurationType
) -> RusotoFuture<(), CreateLaunchConfigurationError>

Creates a launch configuration.

If you exceed your maximum limit of launch configurations, the call fails. For information about viewing this limit, see DescribeAccountLimits. For information about updating this limit, see Amazon EC2 Auto Scaling Limits in the Amazon EC2 Auto Scaling User Guide.

For more information, see Launch Configurations in the Amazon EC2 Auto Scaling User Guide.

fn create_or_update_tags(
&self,
input: CreateOrUpdateTagsType
) -> RusotoFuture<(), CreateOrUpdateTagsError>

Creates or updates tags for the specified Auto Scaling group.

When you specify a tag with a key that already exists, the operation overwrites the previous tag definition, and you do not get an error message.

For more information, see Tagging Auto Scaling Groups and Instances in the Amazon EC2 Auto Scaling User Guide.

fn delete_auto_scaling_group(
&self,
input: DeleteAutoScalingGroupType
) -> RusotoFuture<(), DeleteAutoScalingGroupError>

Deletes the specified Auto Scaling group.

If the group has instances or scaling activities in progress, you must specify the option to force the deletion in order for it to succeed.

If the group has policies, deleting the group deletes the policies, the underlying alarm actions, and any alarm that no longer has an associated action.

To remove instances from the Auto Scaling group before deleting it, call DetachInstances with the list of instances and the option to decrement the desired capacity. This ensures that Amazon EC2 Auto Scaling does not launch replacement instances.

To terminate all instances before deleting the Auto Scaling group, call UpdateAutoScalingGroup and set the minimum size and desired capacity of the Auto Scaling group to zero.

fn delete_launch_configuration(
&self,
input: LaunchConfigurationNameType
) -> RusotoFuture<(), DeleteLaunchConfigurationError>

Deletes the specified launch configuration.

The launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.

fn delete_lifecycle_hook(
&self,
input: DeleteLifecycleHookType
) -> RusotoFuture<DeleteLifecycleHookAnswer, DeleteLifecycleHookError>

Deletes the specified lifecycle hook.

If there are any outstanding lifecycle actions, they are completed first (ABANDON for launching instances, CONTINUE for terminating instances).

fn delete_notification_configuration(
&self,
input: DeleteNotificationConfigurationType
) -> RusotoFuture<(), DeleteNotificationConfigurationError>

Deletes the specified notification.

fn delete_policy(
&self,
input: DeletePolicyType
) -> RusotoFuture<(), DeletePolicyError>

Deletes the specified scaling policy.

Deleting a policy deletes the underlying alarm action, but does not delete the alarm, even if it no longer has an associated action.

fn delete_scheduled_action(
&self,
input: DeleteScheduledActionType
) -> RusotoFuture<(), DeleteScheduledActionError>

Deletes the specified scheduled action.

fn delete_tags(
&self,
input: DeleteTagsType
) -> RusotoFuture<(), DeleteTagsError>

Deletes the specified tags.

fn describe_account_limits(
&self
) -> RusotoFuture<DescribeAccountLimitsAnswer, DescribeAccountLimitsError>

Describes the current Amazon EC2 Auto Scaling resource limits for your AWS account.

For information about requesting an increase in these limits, see Amazon EC2 Auto Scaling Limits in the Amazon EC2 Auto Scaling User Guide.

fn describe_adjustment_types(
&self
) -> RusotoFuture<DescribeAdjustmentTypesAnswer, DescribeAdjustmentTypesError>

Describes the policy adjustment types for use with PutScalingPolicy.

fn describe_auto_scaling_groups(
&self,
input: AutoScalingGroupNamesType
) -> RusotoFuture<AutoScalingGroupsType, DescribeAutoScalingGroupsError>

Describes one or more Auto Scaling groups.

fn describe_auto_scaling_instances(
&self,
input: DescribeAutoScalingInstancesType
) -> RusotoFuture<AutoScalingInstancesType, DescribeAutoScalingInstancesError>

Describes one or more Auto Scaling instances.

fn describe_auto_scaling_notification_types(
&self
) -> RusotoFuture<DescribeAutoScalingNotificationTypesAnswer, DescribeAutoScalingNotificationTypesError>

Describes the notification types that are supported by Amazon EC2 Auto Scaling.

fn describe_launch_configurations(
&self,
input: LaunchConfigurationNamesType
) -> RusotoFuture<LaunchConfigurationsType, DescribeLaunchConfigurationsError>

Describes one or more launch configurations.

fn describe_lifecycle_hook_types(
&self
) -> RusotoFuture<DescribeLifecycleHookTypesAnswer, DescribeLifecycleHookTypesError>

Describes the available types of lifecycle hooks.

The following hook types are supported:

autoscaling:EC2INSTANCELAUNCHING

autoscaling:EC2INSTANCETERMINATING

fn describe_lifecycle_hooks(
&self,
input: DescribeLifecycleHooksType
) -> RusotoFuture<DescribeLifecycleHooksAnswer, DescribeLifecycleHooksError>

Describes the lifecycle hooks for the specified Auto Scaling group.

fn describe_load_balancer_target_groups(
&self,
input: DescribeLoadBalancerTargetGroupsRequest
) -> RusotoFuture<DescribeLoadBalancerTargetGroupsResponse, DescribeLoadBalancerTargetGroupsError>

Describes the target groups for the specified Auto Scaling group.

fn describe_load_balancers(
&self,
input: DescribeLoadBalancersRequest
) -> RusotoFuture<DescribeLoadBalancersResponse, DescribeLoadBalancersError>

Describes the load balancers for the specified Auto Scaling group.

This operation describes only Classic Load Balancers. If you have Application Load Balancers or Network Load Balancers, use DescribeLoadBalancerTargetGroups instead.

fn describe_metric_collection_types(
&self
) -> RusotoFuture<DescribeMetricCollectionTypesAnswer, DescribeMetricCollectionTypesError>

Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling.

The GroupStandbyInstances metric is not returned by default. You must explicitly request this metric when calling EnableMetricsCollection.

fn describe_notification_configurations(
&self,
input: DescribeNotificationConfigurationsType
) -> RusotoFuture<DescribeNotificationConfigurationsAnswer, DescribeNotificationConfigurationsError>

Describes the notification actions associated with the specified Auto Scaling group.

fn describe_policies(
&self,
input: DescribePoliciesType
) -> RusotoFuture<PoliciesType, DescribePoliciesError>

Describes the policies for the specified Auto Scaling group.

fn describe_scaling_activities(
&self,
input: DescribeScalingActivitiesType
) -> RusotoFuture<ActivitiesType, DescribeScalingActivitiesError>

Describes one or more scaling activities for the specified Auto Scaling group.

fn describe_scaling_process_types(
&self
) -> RusotoFuture<ProcessesType, DescribeScalingProcessTypesError>

Describes the scaling process types for use with ResumeProcesses and SuspendProcesses.

fn describe_scheduled_actions(
&self,
input: DescribeScheduledActionsType
) -> RusotoFuture<ScheduledActionsType, DescribeScheduledActionsError>

Describes the actions scheduled for your Auto Scaling group that haven't run. To describe the actions that have already run, use DescribeScalingActivities.

fn describe_tags(
&self,
input: DescribeTagsType
) -> RusotoFuture<TagsType, DescribeTagsError>

Describes the specified tags.

You can use filters to limit the results. For example, you can query for the tags for a specific Auto Scaling group. You can specify multiple values for a filter. A tag must match at least one of the specified values for it to be included in the results.

You can also specify multiple filters. The result includes information for a particular tag only if it matches all the filters. If there's no match, no special message is returned.

fn describe_termination_policy_types(
&self
) -> RusotoFuture<DescribeTerminationPolicyTypesAnswer, DescribeTerminationPolicyTypesError>

Describes the termination policies supported by Amazon EC2 Auto Scaling.

For more information, see Controlling Which Auto Scaling Instances Terminate During Scale In in the Amazon EC2 Auto Scaling User Guide.

fn detach_instances(
&self,
input: DetachInstancesQuery
) -> RusotoFuture<DetachInstancesAnswer, DetachInstancesError>

Removes one or more instances from the specified Auto Scaling group.

After the instances are detached, you can manage them independent of the Auto Scaling group.

If you do not specify the option to decrement the desired capacity, Amazon EC2 Auto Scaling launches instances to replace the ones that are detached.

If there is a Classic Load Balancer attached to the Auto Scaling group, the instances are deregistered from the load balancer. If there are target groups attached to the Auto Scaling group, the instances are deregistered from the target groups.

For more information, see Detach EC2 Instances from Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn detach_load_balancer_target_groups(
&self,
input: DetachLoadBalancerTargetGroupsType
) -> RusotoFuture<DetachLoadBalancerTargetGroupsResultType, DetachLoadBalancerTargetGroupsError>

Detaches one or more target groups from the specified Auto Scaling group.

fn detach_load_balancers(
&self,
input: DetachLoadBalancersType
) -> RusotoFuture<DetachLoadBalancersResultType, DetachLoadBalancersError>

Detaches one or more Classic Load Balancers from the specified Auto Scaling group.

This operation detaches only Classic Load Balancers. If you have Application Load Balancers or Network Load Balancers, use DetachLoadBalancerTargetGroups instead.

When you detach a load balancer, it enters the Removing state while deregistering the instances in the group. When all instances are deregistered, then you can no longer describe the load balancer using DescribeLoadBalancers. The instances remain running.

fn disable_metrics_collection(
&self,
input: DisableMetricsCollectionQuery
) -> RusotoFuture<(), DisableMetricsCollectionError>

Disables group metrics for the specified Auto Scaling group.

fn enable_metrics_collection(
&self,
input: EnableMetricsCollectionQuery
) -> RusotoFuture<(), EnableMetricsCollectionError>

Enables group metrics for the specified Auto Scaling group. For more information, see Monitoring Your Auto Scaling Groups and Instances in the Amazon EC2 Auto Scaling User Guide.

fn enter_standby(
&self,
input: EnterStandbyQuery
) -> RusotoFuture<EnterStandbyAnswer, EnterStandbyError>

Moves the specified instances into the standby state.

For more information, see Temporarily Removing Instances from Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn execute_policy(
&self,
input: ExecutePolicyType
) -> RusotoFuture<(), ExecutePolicyError>

Executes the specified policy.

fn exit_standby(
&self,
input: ExitStandbyQuery
) -> RusotoFuture<ExitStandbyAnswer, ExitStandbyError>

Moves the specified instances out of the standby state.

For more information, see Temporarily Removing Instances from Your Auto Scaling Group in the Amazon EC2 Auto Scaling User Guide.

fn put_lifecycle_hook(
&self,
input: PutLifecycleHookType
) -> RusotoFuture<PutLifecycleHookAnswer, PutLifecycleHookError>

Creates or updates a lifecycle hook for the specified Auto Scaling group.

A lifecycle hook tells Amazon EC2 Auto Scaling to perform an action on an instance when the instance launches (before it is put into service) or as the instance terminates (before it is fully terminated).

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state using using RecordLifecycleActionHeartbeat.

If you finish before the timeout period ends, complete the lifecycle action using CompleteLifecycleAction.

For more information, see Amazon EC2 Auto Scaling Lifecycle Hooks in the Amazon EC2 Auto Scaling User Guide.

If you exceed your maximum limit of lifecycle hooks, which by default is 50 per Auto Scaling group, the call fails.

You can view the lifecycle hooks for an Auto Scaling group using DescribeLifecycleHooks. If you are no longer using a lifecycle hook, you can delete it using DeleteLifecycleHook.

fn put_notification_configuration(
&self,
input: PutNotificationConfigurationType
) -> RusotoFuture<(), PutNotificationConfigurationError>

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to the specified topic can have messages delivered to an endpoint such as a web server or an email address.

This configuration overwrites any existing configuration.

For more information, see Getting Amazon SNS Notifications When Your Auto Scaling Group Scales in the Amazon EC2 Auto Scaling User Guide.

fn put_scaling_policy(
&self,
input: PutScalingPolicyType
) -> RusotoFuture<PolicyARNType, PutScalingPolicyError>

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameters to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request.

fn put_scheduled_update_group_action(
&self,
input: PutScheduledUpdateGroupActionType
) -> RusotoFuture<(), PutScheduledUpdateGroupActionError>

Creates or updates a scheduled scaling action for an Auto Scaling group. If you leave a parameter unspecified when updating a scheduled scaling action, the corresponding value remains unchanged.

For more information, see Scheduled Scaling in the Amazon EC2 Auto Scaling User Guide.

fn record_lifecycle_action_heartbeat(
&self,
input: RecordLifecycleActionHeartbeatType
) -> RusotoFuture<RecordLifecycleActionHeartbeatAnswer, RecordLifecycleActionHeartbeatError>

Records a heartbeat for the lifecycle action associated with the specified token or instance. This extends the timeout by the length of time defined using PutLifecycleHook.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle in the Amazon EC2 Auto Scaling User Guide.

fn resume_processes(
&self,
input: ScalingProcessQuery
) -> RusotoFuture<(), ResumeProcessesError>

Resumes the specified suspended automatic scaling processes, or all suspended process, for the specified Auto Scaling group.

For more information, see Suspending and Resuming Scaling Processes in the Amazon EC2 Auto Scaling User Guide.

fn set_desired_capacity(
&self,
input: SetDesiredCapacityType
) -> RusotoFuture<(), SetDesiredCapacityError>

Sets the size of the specified Auto Scaling group.

For more information about desired capacity, see What Is Amazon EC2 Auto Scaling? in the Amazon EC2 Auto Scaling User Guide.

fn set_instance_health(
&self,
input: SetInstanceHealthQuery
) -> RusotoFuture<(), SetInstanceHealthError>

Sets the health status of the specified instance.

For more information, see Health Checks for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

fn set_instance_protection(
&self,
input: SetInstanceProtectionQuery
) -> RusotoFuture<SetInstanceProtectionAnswer, SetInstanceProtectionError>

Updates the instance protection settings of the specified instances.

For more information about preventing instances that are part of an Auto Scaling group from terminating on scale in, see Instance Protection in the Amazon EC2 Auto Scaling User Guide.

fn suspend_processes(
&self,
input: ScalingProcessQuery
) -> RusotoFuture<(), SuspendProcessesError>

Suspends the specified automatic scaling processes, or all processes, for the specified Auto Scaling group.

If you suspend either the Launch or Terminate process types, it can prevent other process types from functioning properly.

To resume processes that have been suspended, use ResumeProcesses.

For more information, see Suspending and Resuming Scaling Processes in the Amazon EC2 Auto Scaling User Guide.

fn terminate_instance_in_auto_scaling_group(
&self,
input: TerminateInstanceInAutoScalingGroupType
) -> RusotoFuture<ActivityType, TerminateInstanceInAutoScalingGroupError>

Terminates the specified instance and optionally adjusts the desired group size.

This call simply makes a termination request. The instance is not terminated immediately.

fn update_auto_scaling_group(
&self,
input: UpdateAutoScalingGroupType
) -> RusotoFuture<(), UpdateAutoScalingGroupError>

Updates the configuration for the specified Auto Scaling group.

The new settings take effect on any scaling activities after this call returns. Scaling activities that are currently in progress aren't affected.

To update an Auto Scaling group with a launch configuration with InstanceMonitoring set to false, you must first disable the collection of group metrics. Otherwise, you get an error. If you have previously enabled the collection of group metrics, you can disable it using DisableMetricsCollection.

Note the following:

If you specify a new value for MinSize without specifying a value for DesiredCapacity, and the new MinSize is larger than the current size of the group, we implicitly call SetDesiredCapacity to set the size of the group to the new value of MinSize.

If you specify a new value for MaxSize without specifying a value for DesiredCapacity, and the new MaxSize is smaller than the current size of the group, we implicitly call SetDesiredCapacity to set the size of the group to the new value of MaxSize.

All other optional parameters are left unchanged if not specified.

Loading content...